@@ -1,9 +1,8 @@
-﻿namespace VideoProcessor
+﻿namespace VideoProcessor;
+
+public class Approval
 {
-    public class Approval
-    {
-        public string PartitionKey { get; set; }
-        public string RowKey { get; set; }
-        public string OrchestrationId { get; set; }
-    }
+    public string PartitionKey { get; set; }
+    public string RowKey { get; set; }
+    public string OrchestrationId { get; set; }
 }
