@@ -1,8 +1,7 @@
-﻿namespace VideoProcessor
+﻿namespace VideoProcessor;
+
+public class VideoFileInfo
 {
-    public class VideoFileInfo
-    {
-        public string Location { get; set; }
-        public int BitRate { get; set; }
-    }
+    public string Location { get; set; }
+    public int BitRate { get; set; }
 }
